Factor out routes.ts
