+++ arches_lingo/src/arches_lingo/pages/HomePage.vue
@@ -6,13 +6,13 @@ import { useGettext } from "vue3-gettext";
 import { useToast } from "primevue/usetoast";
 import Button from "primevue/button";
 
+import { logout } from "@/arches_lingo/api.ts";
 import {
     DEFAULT_ERROR_TOAST_LIFE,
     ERROR,
-    routeNames,
+    userKey,
 } from "@/arches_lingo/constants.ts";
-import { logout } from "@/arches_lingo/api.ts";
-import { userKey } from "@/arches_lingo/constants.ts";
+import { routeNames } from "@/arches_lingo/routes.ts";
 
 import type { UserRefAndSetter } from "@/arches_lingo/types";
